server/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
cloudinary==1.40.0
coreschema==0.0.4
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
